move properties from measures -> utils
